requirements.txt
astropy==3.2.1
cycler==0.10.0
healpy==1.12.10
importlib-resources==1.0.2
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
pandas==0.25.1
-e git+git@github.com:cjm96/NullStreams.git@5fe7e576ca4e280b5f77cb16c8292efc8b7701c3#egg=ptacake
pyparsing==2.4.2
python-dateutil==2.8.0
pytz==2019.2
scipy==1.3.1
six==1.12.0
